employeeproject/employeeproject/requirements.txt
Django==3.0.3
djangorestframework==3.11.1
requests==2.24.0
django-rest-swagger==2.2.0
mysqlclient==2.0.1
django-cors-headers==3.4.0
whitenoise==5.2.0
python-dateutil==2.4.1
django-request-logging==0.7.2
six==1.15.0
django-rest-passwordreset==1.1.0
django-email-verification==0.0.7
pyotp==2.4.1
bcrypt==3.2.0
django-smtp-ssl==1.0
celery[sqs]==5.0.4
django-debug-toolbar==3.0
django-debug-toolbar-request-history==0.1.3
stripe==2.50.0
cursor==1.3.4
cryptography==3.4.6
djangorestframework-simplejwt==4.6.0
pyjwt==2.1.0
django-smtp-ssl==1.0
django-email-verification==0.0.7
drf-yasg==1.20.0
gunicorn==20.1.0

